+++ Makefile
@@ -59,10 +59,10 @@ help: ## Display this help.
 	@awk 'BEGIN {FS = ":.*##"; printf "\nUsage:\n  make \033[36m<target>\033[0m\n"} /^[a-zA-Z_0-9-]+:.*?##/ { printf "  \033[36m%-15s\033[0m %s\n", $$1, $$2 } /^##@/ { printf "\n\033[1m%s\033[0m\n", substr($$0, 5) } ' $(MAKEFILE_LIST)
 
 .PHONY: static
-static: $(TAILWIND_JS) manifests generate fmt ginkgo ui
+static: manifests generate fmt ginkgo ui
 
 .PHONY: ui
-ui: $(LOCALBIN)/pnpm ## Build the embedded catalog explorer UI (ui/frontend -> ui/frontend/dist/ui.js)
+ui: $(LOCALBIN)/pnpm $(TAILWIND_JS) ## Build the embedded catalog explorer UI (ui/frontend -> ui/frontend/dist/ui.js)
 	cd ui/frontend && CI=true pnpm install --no-frozen-lockfile --prefer-offline && pnpm run build
 
 .PHONY: plugins-ui
