+++ plugins/kubernetes-logs/main.go
@@ -9,14 +9,9 @@ package main
 import (
 	"context"
 	"embed"
-	"encoding/json"
-	"fmt"
 	"io/fs"
 	"net/http"
 
-	dutyContext "github.com/flanksource/duty/context"
-	dutylogs "github.com/flanksource/duty/logs"
-	v1 "github.com/flanksource/incident-commander/api/v1"
 	pluginpb "github.com/flanksource/incident-commander/plugin/proto"
 	"github.com/flanksource/incident-commander/plugin/sdk"
 )
@@ -96,81 +91,4 @@ func (p *KubernetesLogsPlugin) Operations() []sdk.Operation {
 			HTTPHandler: http.HandlerFunc(p.httpLogs),
 		},
 	}
-}
-
-// TailParams is the input shape for the `tail` operation. The CLI sends
-// these as JSON via --param/--json; the iframe sends them through the
-// `tail` button. PostProcess mirrors the playbook `logs` action shape so
-// CEL match expressions and dedup/window settings work the same way.
-type TailParams struct {
-	Container   string             `json:"container,omitempty"`
-	TailLines   int64              `json:"tailLines,omitempty"`
-	Previous    bool               `json:"previous,omitempty"`
-	PostProcess v1.LogsPostProcess `json:"postProcess,omitempty"`
-}
-
-func (p *KubernetesLogsPlugin) tail(ctx context.Context, req sdk.InvokeCtx) (any, error) {
-	var params TailParams
-	if len(req.ParamsJSON) > 0 {
-		if err := json.Unmarshal(req.ParamsJSON, &params); err != nil {
-			return nil, fmt.Errorf("decode params: %w", err)
-		}
-	}
-	if params.TailLines <= 0 {
-		params.TailLines = 200
-	}
-
-	cli, err := p.clients.For(ctx, req.Host, req.ConfigItemID)
-	if err != nil {
-		return nil, err
-	}
-
-	pods, err := resolvePods(ctx, cli, req.Host, req.ConfigItemID)
-	if err != nil {
-		return nil, err
-	}
-
-	var lines []*dutylogs.LogLine
-	for _, pod := range pods {
-		podLines, err := tailPod(ctx, cli, pod, params)
-		if err != nil {
-			lines = append(lines, errorLine(pod, "", err.Error()))
-			continue
-		}
-		lines = append(lines, podLines...)
-	}
-
-	dctx := dutyContext.NewContext(ctx)
-	return postProcessLogs(dctx, lines, params.PostProcess), nil
-}
-
-func (p *KubernetesLogsPlugin) listPods(ctx context.Context, req sdk.InvokeCtx) (any, error) {
-	cli, err := p.clients.For(ctx, req.Host, req.ConfigItemID)
-	if err != nil {
-		return nil, err
-	}
-	pods, err := resolvePods(ctx, cli, req.Host, req.ConfigItemID)
-	if err != nil {
-		return nil, err
-	}
-	type Row struct {
-		Namespace string `json:"namespace"`
-		Pod       string `json:"pod"`
-		Phase     string `json:"phase"`
-		OwnedBy   string `json:"ownedBy,omitempty"`
-	}
-	out := make([]Row, 0, len(pods))
-	for _, pod := range pods {
-		owned := ""
-		if len(pod.OwnerReferences) > 0 {
-			owned = fmt.Sprintf("%s/%s", pod.OwnerReferences[0].Kind, pod.OwnerReferences[0].Name)
-		}
-		out = append(out, Row{
-			Namespace: pod.Namespace,
-			Pod:       pod.Name,
-			Phase:     string(pod.Status.Phase),
-			OwnedBy:   owned,
-		})
-	}
-	return out, nil
 }
